Split System\Security\Attributes.cs into properly named files.
