+++ src/mscorlib/shared/System/Security/SuppressUnmanagedCodeSecurityAttribute.cs
@@ -1,0 +1,15 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Security
+{
+    // SuppressUnmanagedCodeSecurityAttribute:
+    //  Indicates that the target P/Invoke method(s) should skip the per-call
+    //  security checked for unmanaged code permission.
+    [AttributeUsage(AttributeTargets.Method | AttributeTargets.Class | AttributeTargets.Interface | AttributeTargets.Delegate, AllowMultiple = true, Inherited = false)]
+    sealed public class SuppressUnmanagedCodeSecurityAttribute : System.Attribute
+    {
+    }
+}
+
